fixed_quantity semantics unit tests fixup now ratinal has a long long integer type

+++ quan_matters/test/fixed_quantity/semantics/dimless_multiply.cpp
@@ -269,7 +269,10 @@ namespace{
        typedef quan::detail::dimensionless_multiply1<
            conversion_factor_L, conversion_factor_R
        > mul;
-       QUAN_CHECK((typeid(mul::eval<int,int>()(1,1)) == typeid(int)));
+//#########################promoted#########################
+       QUAN_CHECK((typeid(mul::eval<int,int>()(1,1)) == typeid(long long)));
+//#########################################################################
+    //    std::cout << typeid(mul::eval<int,int>()(1,1)).name() <<'\n';
        int val1 = mul::eval<int,int>()(1,1);
        int val1_chk = (1 * 1) * ((7 / 1) * (9 / 1));
        QUAN_CHECK_EQUAL(val1,val1_chk);
@@ -617,7 +620,9 @@ namespace{
            conversion_factor_R::multiplier
        >::type rat;
      //  std::cout << rat::numerator << "/ " << rat::denominator <<'\n';
-       QUAN_CHECK( (typeid(mul::eval<int,int>()(1,1)) == typeid(int)));
+//##################modified######################
+       QUAN_CHECK( (typeid(mul::eval<int,int>()(1,1)) == typeid(long long)));
+//###########################################
        double in_L = 2.8f;
        double in_R = 7.2f;
        double result = mul::eval<double,double>()(in_L,in_R);
